corrected en_passant

+++ rust_compute/src/chess_board.rs
@@ -771,7 +771,7 @@ impl ChessBoard {
             target_black[i] = biases[i] as i16;
         }
 
-        // FIX 1: Combine ALL piece bitboards to ensure nothing (especially the Kings!) is missed.
+        // Combine ALL piece bitboards to ensure nothing (especially the Kings!) is missed.
         let mut pieces_bitboard = (self.pawns[0] | self.pawns[1])
 
             | (self.knights[0] | self.knights[1])
@@ -803,7 +803,7 @@ impl ChessBoard {
         // --- LOOP 1: PROCESS WHITE PERSPECTIVE COMPLETELY ---
         // Perfect auto-vectorization! The CPU fills all registers exclusively with White data.
         for &w_idx in white_indices.iter().take(active_piece_count) {
-            let w_row = &self.nnue_network.l1_weights[w_idx as usize][..256];
+            let w_row = &self.nnue_network.l1_weights[w_idx][..256];
             
             for i in 0..256 {
                 target_white[i] = target_white[i].wrapping_add(w_row[i]);
@@ -813,7 +813,7 @@ impl ChessBoard {
         // --- LOOP 2: PROCESS BLACK PERSPECTIVE COMPLETELY ---
         // Perfect auto-vectorization! The CPU reuses those same registers exclusively for Black data.
         for &b_idx in black_indices.iter().take(active_piece_count) {
-            let b_row = &self.nnue_network.l1_weights[b_idx as usize][..256];
+            let b_row = &self.nnue_network.l1_weights[b_idx][..256];
             
             for i in 0..256 {
                 target_black[i] = target_black[i].wrapping_add(b_row[i]);
